move libp2p logic out of default implementation
