+++ crates/hotshot/src/traits/networking/combined_network.rs
@@ -466,7 +466,7 @@ impl<TYPES: NodeType> ConnectedNetwork<Message<TYPES>, TYPES::SignatureKey>
         self.secondary().queue_node_lookup(view_number, pk).await
     }
 
-    async fn update_view<'a, T>(&'a self, view: u64, _membership: &T::Membership)
+    async fn update_view<'a, T>(&'a self, view: u64, membership: &T::Membership)
     where
         T: NodeType<SignatureKey = TYPES::SignatureKey> + 'a,
     {
@@ -494,6 +494,9 @@ impl<TYPES: NodeType> ConnectedNetwork<Message<TYPES>, TYPES::SignatureKey>
         // View changed, let's start primary again
         self.primary_down.store(false, Ordering::Relaxed);
         self.primary_fail_counter.store(0, Ordering::Relaxed);
+
+        // Run `update_view` logic for the libp2p network
+        self.networks.1.update_view::<T>(view, membership).await;
     }
 
     fn is_primary_down(&self) -> bool {
